added error event handler for CORS
My code worked for the past year or so with no problems but lately I've been getting CORS errors when requests are made for words that don't have synonyms in the API; this makes me think the API changed. Regardless, I added an event handler for that case that moves to the next word.
The past many commits have been me changing the code around and trying to fix this error. I know I really shouldn't commit for all those changes but I thought it would be a quick fix and I didn't want to setup the code elsewhere, so I had to just commit changes and see how it changed the github pages page. This is the final commit though.